chore: add delete functionality

+++ model/model.go
@@ -15,7 +15,7 @@ var db *gorm.DB
 
 // Site defines the structure of the db
 type Site struct {
-	ID       uint64
+	ID      int32
 	Name     string
 	UserName string
 	Password string
